% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ar_data}
\alias{ar_data}
\title{A 1\% sample of df_1st_vote}
\format{
A data frame with 69466 rows and 68 variables:
\describe{
\item{\code{id_vot}}{Unique ID of the voting.}
\item{\code{data}}{Data when the voting took place.}
}
}
\source{
{
Data created in-house to use as an example.
Code:
df <- read_ar_data(path = "../Project_data_AR", name = "df_1st_vote+index")
set.seed(12345)
id_vot_select <- df \%>\% distinct(id_vot) \%>\% dplyr::slice_sample(prop = 0.01) \%>\% pull(id_vot)
ar_data <- df \%>\% filter(id_vot \%in\% id_vot_select)
ar_data \%<>\% select(id_vot, data, partido, voto, n_votos, unanime,
id_ati, id_com, id_com_cross, id_ini, tipo_ini,
tipo_atividade, proponente, legis, n_dep, total_dep,
bill_type, vote_stage, votos_legis_partido, media_index)
use_data(ar_data,
compress = "xz",
overwrite = T)
}
}
\usage{
ar_data
}
\description{
A sample of 1 \% of the full cleaned df_1st_vote.
The documentation of the variables is not complete.
}
\examples{
data(ar_data) # lazy loading. The RAM will not be immediately occupied.
}
\keyword{datasets}
