% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ask_chatgpt.R
\name{ask_chatgpt}
\alias{ask_chatgpt}
\title{Calls the ChatGPT API}
\usage{
ask_chatgpt(prompt, api_key = key, cat = TRUE, version = "gpt-3.5-turbo")
}
\arguments{
\item{api_key}{A string. Secret key for chatGPT API account. By default, it's an object named key.}

\item{cat}{Logical value. Apply cat() to the output. TRUE is the default.}

\item{version}{What is the version of chatGPT. "gpt-3.5-turbo" is the default.}

\item{promt}{Text for chat GPT. A string.}
}
\value{
The response of chat GPT.
}
\description{
This function calls the chatGPT API
with the given prompt and returns the answer.
Taken from: https://www.r-bloggers.com/2023/03/call-chatgpt-or-really-any-other-api-from-r/
}
\examples{
# key <- "sk-5-your-actual-api-key-Fvau6"
# ask_chatgpt("Faz-me uma lista de frutas")
# ask_chatgpt("Faz-me uma lista de frutas", cat = F)
}
