% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duration_expand.R
\name{duration_expand}
\alias{duration_expand}
\title{Expand the dates to daily data}
\usage{
duration_expand(data, start, end, group)
}
\arguments{
\item{group}{Group for which to expand}
}
\description{
This file...
}
