% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/media_bill_count.R
\name{media_bill_count}
\alias{media_bill_count}
\title{Count the number of News that talk about a bill.}
\usage{
media_bill_count(
  data,
  var_entity,
  var_date,
  data_media,
  var_media_date,
  var_news,
  var_news_title,
  var_news_id,
  days_treshold = 365,
  ratio_entity_presence = 0.9
)
}
\arguments{
\item{data}{Tibble with entities in bills and their dates}

\item{var_entity}{Name of the variable with the entities. Not a string.}

\item{var_date}{Name of the variable with the dates. Not a string.}

\item{data_media}{Data with the media news}

\item{var_media_date}{Name of the variable with the news time stamp or date.}

\item{var_news}{Name of the variable of the news to use. It can refer to texts, titles, subtitles, etc. Not a string.}

\item{var_news_title}{Name of the varaible with the news titles}

\item{var_news_id}{A unique ID idintifying the news article saved by Arquivo.pt}

\item{days_treshold}{How many days should be given as an interval. The dates of the news are noisy, thus this should be high.}

\item{ratio_entity_presence}{How many (in share) of the bill's entities must be in the same article to be considered a match.}
}
\value{
data with 2 columns, one for extensive and other for extensive plus intensive margin. Plus a vector with all the id of the bills mentioned.
}
\description{
This function takes the entities in the title of bills and counts how many
times they appear in the news.
The function returns both an extensive margin and an intensive plus extensive margin columns.
extensive margin: does a certain entity appear in a bill title?
extensive and intensive margin: how many times does a certain entity appear in a bill title?
The function can parallelize the procedure.
But the number of cores must be set outside using
plan(multisession, workers = x) and with the furrr package loaded.
}
\examples{
# df_gpt \%>\%
#   media_bill_counts(entity,
#                     date,
#                     df_desarquivo,
#                     data,
#                     text,
#                     titulos)


}
