% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_party.R
\name{order_party}
\alias{order_party}
\title{Order the parties}
\usage{
order_party(data, name)
}
\arguments{
\item{data}{a tibble}

\item{name}{Name of the column with party info}
}
\description{
This function orders a factor variable containing parties according to their
Left-right spectrum.
}
\examples{
#data(ar_data)
#ar_data \%>\% order_party(partido)


}
