% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_values.R
\name{p_values}
\alias{p_values}
\title{Calculate p-values from a lm()}
\usage{
p_values(lm_object, correction = NULL, rounded_to = 5)
}
\arguments{
\item{lm_object}{An estimated lm() object}

\item{correction_method}{String. A correction method to pass to p.adjust()}
}
\value{
A named vector with the adjusted p-values.
}
\description{
This function calculates the p-values from a lm() object.
The function can extract the p-values but also allows
for the correction of those p-values, given any multiple testing method
available in the function p-adjust.
Type ?p.adjust.methods for an overvierw.
}
\examples{
# load an example dataset
data(mtcars)

# estimate a linear model
m <- lm(mpg ~ as.factor(cyl) + hp + wt + am, data = mtcars)

# extract the non-corrected p-values
p_values(m)

# extract p-values corrected with the Holm metghod.
m |> p_values(correction = "holm")

}
