% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prox_by_bill.R
\name{prox_by_bill}
\alias{prox_by_bill}
\title{Create a proximity matrix for each bill ####}
\usage{
prox_by_bill(data, type = "tibble", order = "leftright")
}
\arguments{
\item{data}{A tibble or data frame}

\item{type}{A string. Can be "list" or "tibble".}

\item{order}{Organization of parties in the prox matrix. "leftright", the default, organizes parties from left to right. "alphabetical" is also possible.}
}
\value{
A tibble or a list, where each element is a party proximity to a given bill.
}
\description{
This function transforms a tidy dataset with roll-call data into party proximity by each vote.
It outputs a nested tibble or list that shows the Euclidean distance between all party by bill.
If "list" (the default) the function returns a list of matrices.
If type = "tibble", the function returns a tibble where each
line has a vector that is the lower triangular part of a proximity matrix for a bill.
The lower tringular part is organized from top to bottom, from left to right.
The function assigns a value of 1 to an MP that votes in favor, 0 to against and 0.5 to abstentions and misses.
Note: for the boot::boot function, a dataset must be provided.
}
\examples{
library(dplyr)

data("ar_data")
ar_data |>
  filter(legis \%in\% c(14, 15)) \%>\%
  prox_by_bill(type = "tibble")

}
