% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prox_matrix.R
\name{prox_matrix}
\alias{prox_matrix}
\title{Build a Proximity Matrix}
\usage{
prox_matrix(
  dataset,
  ...,
  legislature = c(1:20),
  commission = c(1:11),
  type = c("law", NA, "resolution", "constitutional", "correction", "inquiry",
    "recommendation", "validation", "statute"),
  stage = c("general", "final", "other", "specific", "recommendation"),
  vec_parties = c("CDU", "PCP", "PEV", "ID", "MDP/CDE", "MDP", "BE", "UDP", "L", "PAN",
    "UEDS", "PS", "ASDI", "PSN", "PRD", "PSD", "IL", "CDS", "PPM", "CH"),
  initiated_by = c("Government and others", "CDU", "PCP", "PEV", "ID", "MDP/CDE", "MDP",
    "BE", "UDP", "L", "PAN", "UEDS", "PS", "ASDI", "PSN", "PRD", "PSD", "IL", "CDS",
    "PPM", "CH"),
  min_n_votes = 0,
  unanimous = c(0, 1),
  plot_plotly = TRUE,
  plot = TRUE,
  inference = F,
  n_boot = 1000,
  seed_for_boot = 12345
)
}
\arguments{
\item{dataset}{A tibble}

\item{...}{Parties to consider in the function.}

\item{legislature}{A numeric vector with legislatures to plot.}

\item{commission}{A numeric vector with the ID of the commissions crosswalk.}

\item{type}{Type of law. A character vector.}

\item{stage}{A character vector with the stage in which the vote was taken.}

\item{vec_parties}{It does the same as "...", but the input is a numeric vector. Useful for the shiny app.}

\item{initiated_by}{A character vector with the parties that proposed the law to parliament.}

\item{min_n_votes}{The relationship between two parties is only shown if it meets this minimum requirement.}

\item{unanimous}{If 0, show only non unanimous. If one show only unanimous.}

\item{plot_plotly}{The output is a plotly, not a ggplot matrix.}

\item{plot}{Return the dataset p2, not a plot.}

\item{inference}{Use the bootstrapping inference to build confidence intervals.}

\item{n_boot}{Number of re-sampling for the bootstrap.}

\item{seed_for_boot}{Seed for the bootstap}
}
\value{
A matrix graph.
}
\description{
Build a proximity matrix of the parties.
The function considers agreements to count as "1" and disagreements to count as "-1",
while abstentions count as 0. Then, an average is performed.
If no input is provided, the function will use all available votes in the dataset.
}
\examples{
data(ar_data)

# ar_data |> prox_matrix()
ar_data |> prox_matrix(legislature = 15)

}
