% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ar_data.R
\name{read_ar_data}
\alias{read_ar_data}
\title{Read AR Data}
\usage{
read_ar_data(name = NULL, path = NULL)
}
\arguments{
\item{name}{Name of the dataset to load. A string.}

\item{path}{Path to the folder "Project_data_AR", if the function is not ran within the project.}
}
\value{
A dataset or a list of available datasets, if "name" is left in blank.
}
\description{
Read cleaned data with party votes from the Portuguese parliament.
}
\examples{
# to list available data
#read_ar_data()
read_ar_data(path = "../Project_data_AR")

# to load data
# df <- read_ar_data("df_1st_vote")


}
