% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stretch_for_df_unique.R
\name{stretch_for_df_unique}
\alias{stretch_for_df_unique}
\title{From lower triangular to full}
\usage{
stretch_for_df_unique(x, diag_to_insert = NULL)
}
\arguments{
\item{x}{a vectror. Lower tringualar matrix as a vector.}

\item{diag_to_insert}{The diagonal to insert. By default is 1 to all elements.}
}
\value{
A vector as a full proximity matrix from top to bottom, from left to right.
}
\description{
The output from politicR::prox_by_bill and boot_prox is the
lower triangular part of a proximity matrix, from top to bottom, from left to right.
This function transforms it into a vector, of a full proximity matrix,
from top to bottom, from left to right.
}
