% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{variance}
\alias{variance}
\title{Compute Population Variance}
\usage{
variance(x, na.rm = TRUE, single_value = TRUE)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{A logical value indicating whether \code{NA} values should be removed before the computation.
The default is \code{TRUE}.}

\item{single_value}{Instead of attributing \code{NA} to single value vectors, atributes 0.}
}
\value{
A numeric value representing the population variance of the input vector.
}
\description{
This function calculates the population variance of a numeric vector, with an option to handle missing values.
}
\details{
The function adjusts the sample variance (calculated by \code{var()}) by multiplying it by \((n-1)/n\),
where \(n\) is the length of the input vector, to derive the population variance.
}
\examples{
# Example usage
data <- c(1, 2, 3, 4, 5)
variance(data)

data_with_na <- c(1, 2, 3, NA, 5)
variance(data_with_na, na.rm = TRUE)
variance(data_with_na, na.rm = FALSE)

}
