% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_salience.R
\name{weight_salience}
\alias{weight_salience}
\title{Weight bills by their importance.}
\usage{
weight_salience(vec_prox, vec_media_index)
}
\arguments{
\item{vec_prox}{A vector with the proximity between 2 parties over a set of bills.}

\item{vec_media_index}{A vector with the respective importance of those bills.}
}
\value{
The proximity of the parties weighted by the importance.
}
\description{
With just two inputs, it creates a weighted average with normalization.
}
\examples{
# The proximities between two parties over 5 bills
vec_proximity <- c(0.1, 0.5, 0.6, 0.4, 0.9)

# the media index of each bill
vec_media_indexes <- c(5, 40, 30, 70, 10)

weight_salience(vec_proximity,
                vec_media_indexes)



}
