% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_prox.R
\name{boot_prox}
\alias{boot_prox}
\title{Bootstap Funtion for Party Proximity}
\usage{
boot_prox(data, indices = NULL)
}
\arguments{
\item{data}{A tibble where each line is a bill, and has a proximity vector nested,}

\item{indices}{The indexes to be retured. Necessary for the bootstrap.}
}
\value{
A vector with the
}
\description{
Function to use in the bootstraping.
}
\examples{
#ta bem

}
