% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prox_by_bill.R
\name{prox_by_bill}
\alias{prox_by_bill}
\title{Create a proximity matrix for each bill ####}
\usage{
prox_by_bill(data, type = "tibble", order = "leftright")
}
\arguments{
\item{data}{A tibble or data frame}

\item{type}{A string. Can be "list" or "tibble".}
}
\value{
A tibble or a matrix, where each element is a vote containing the party proximity per bill. But the object is streched.
}
\description{
Eucledian distance. Ina 2d pane, it's the absolute difference.
If "list" (the default) the function returns a list of matrices, where each
elements is a bill. If "tibble", the function returns a tibble where each
line has another tibble as matrix.
Note: for the boot::boot function, a dataset must be provided.
But the object returned it streched. The function prox_matrix(2 for now) to transform it into a matrix.
}
\examples{
library(dplyr)

data("ar_data")
ar_data |>
  filter(legis \%in\% c(14, 15)) \%>\%
  prox_by_bill(type = "tibble")

}
